Refactored tasks and projects
Moved both tasks and project logic into the index js
to allow for simpler passing of props and page management.
