+++ src/global-style.css
@@ -8,3 +8,4 @@ body, p, h2, h3 {
 .hidden {
   display: none !important;
 }
+
